Add script for modifying .mozbuild sysroot
As noted in cpsysroot.sh, we have a mismatch between the version of libc that is used to build partition-alloc and the version in Firefox's custom sysroot. This results in linker errors because Firefox's libc doesn't have all of the symbols that PA requires. To work around this, we copy the necessary system libraries into the sysroot.

This also required some minor header changes. In some cases, the build system tests for the presence of libc symbols and sets preprocessor defines based on the presence of some symbols. In ptthread.c it attempts to use gettid, which isn't present in the default sysroot but is present in the newer system library. This causes a mismatch because even though the symbol is present in the updated libc, we're still using the old libc headers which do not define that symbol. To work around this, I've modified the header to always define the custom gettid that is used normally.

A similar situation occurs in `event-config.h`, where the build system checks for the presence of some arc4random functions. The symbols are present, so the build system sets corresponding defines, but then the headers do not define the functions. I have hard coded it to assume that the symbols are not present so that it uses the shim arc4random library that it normally uses when we don't modify the sysroot.