Put environment variables on the shared heap
Firefox sets some environment variables at runtime, which means some of the env vars point to compartment 1's heap. libjpeg then tries to read env vars, causing a compartment violation when it gets the vars that were set at runtime. To fix this we copy any strings passed to `setenv` and `putenv` to the shared heap before setting the env var. This leaks some memory, but probably not much.