Move jpeg buffers to shared heap
These are the remaining allocations that libjpeg needs access to.

One minor note: We change `SourceBuffer::Chunk` and `DownscalingFilter` to always allocate their data on the shared heap, but I'm pretty sure these are used for other decoding purposes beyond just libjpeg. Doing it this way is potentially a security hole if libjpeg somehow gets control of a buffer for a different decoding process. A more robust solution might be to only have chunks/filters that are used with libjpeg be allocated on the shared heap, but that's likely a more involved change that isn't worth doing for the purpose of the Firefox demo.